Moving into separate folders per tool
