+++ src/tools/tools.ts
@@ -1,9 +1,9 @@
 import { listDatasourcesTool } from './listDatasources/listDatasources.js';
 import { listFieldsTool } from './listFields.js';
-import { listAllPulseMetricDefinitionsTool } from './pulse/listMetricDefinitions/listAllPulseMetricDefinitions.js';
-import { listPulseMetricDefinitionsFromDefinitionIdsTool } from './pulse/listMetricDefinitions/listPulseMetricDefinitionsFromDefinitionIds.js';
-import { listPulseMetricsFromMetricDefinitionIdTool } from './pulse/listMetrics/listPulseMetricsFromMetricDefinitionId.js';
-import { listPulseMetricsFromMetricIdsTool } from './pulse/listMetrics/listPulseMetricsFromMetricIds.js';
+import { listAllPulseMetricDefinitionsTool } from './pulse/listAllMetricDefinitions/listAllPulseMetricDefinitions.js';
+import { listPulseMetricDefinitionsFromDefinitionIdsTool } from './pulse/listMetricDefinitionsFromDefinitionIds/listPulseMetricDefinitionsFromDefinitionIds.js';
+import { listPulseMetricsFromMetricDefinitionIdTool } from './pulse/listMetricsFromMetricDefinitionId/listPulseMetricsFromMetricDefinitionId.js';
+import { listPulseMetricsFromMetricIdsTool } from './pulse/listMetricsFromMetricIds/listPulseMetricsFromMetricIds.js';
 import { listPulseMetricSubscriptionsTool } from './pulse/listMetricSubscriptions/listPulseMetricSubscriptions.js';
 import { queryDatasourceTool } from './queryDatasource/queryDatasource.js';
 import { readMetadataTool } from './readMetadata.js';
